config is now package with test independed of filesystem
