+++ main.go
@@ -1,29 +1,23 @@
 package main
 
 import (
-	"encoding/json"
 	"github.com/op/go-logging"
-	"io/ioutil"
+	"github.com/gempir/relaybroker/config"
 	"os"
 )
 
-type Config struct {
-	Broker_port string `json:"broker_port"`
-	Broker_pass string `json:"broker_pass"`
-}
-
 var (
-	config Config
+	cfg config.Config
 	log    logging.Logger
 )
 
 func main() {
 	log = initLogger()
-	config, err := readConfig("config.json")
+	cfg, err := config.ReadConfig("config.json")
 	if err != nil {
 		log.Fatal(err)
 	}
-	log.Info("starting up on port", config.Broker_port)
+	log.Info("starting up on port", cfg.Broker_port)
 }
 
 func initLogger() logging.Logger {
@@ -37,17 +31,4 @@ func initLogger() logging.Logger {
 	backend1Leveled.SetLevel(logging.ERROR, "")
 	logging.SetBackend(backend1Leveled, backend2Formatter)
 	return *logger
-}
-
-func readConfig(path string) (Config, error) {
-	var cfg Config
-	file, err := ioutil.ReadFile(path)
-	if err != nil {
-		return cfg, err
-	}
-	err = json.Unmarshal(file, &cfg)
-	if err != nil {
-		return cfg, err
-	}
-	return cfg, nil
 }
